extract the tests out of `methods.rs`
Among other things, this duplicates the repro for issue10029 for `.next_back()`
