+++ tests/ui/filter_next.stderr
@@ -1,5 +1,5 @@
 error: called `filter(..).next()` on an `Iterator`. This is more succinctly expressed by calling `.find(..)` instead
-  --> tests/ui/methods_fixable.rs:9:13
+  --> tests/ui/filter_next.rs:16:13
    |
 LL |     let _ = v.iter().filter(|&x| *x < 0).next();
    |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `v.iter().find(|&x| *x < 0)`
@@ -8,13 +8,13 @@ LL |     let _ = v.iter().filter(|&x| *x < 0).next();
    = help: to override `-D warnings` add `#[allow(clippy::filter_next)]`
 
 error: called `filter(..).next_back()` on an `DoubleEndedIterator`. This is more succinctly expressed by calling `.rfind(..)` instead
-  --> tests/ui/methods_fixable.rs:12:13
+  --> tests/ui/filter_next.rs:19:13
    |
 LL |     let _ = v.iter().filter(|&x| *x < 0).next_back();
    |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `v.iter().rfind(|&x| *x < 0)`
 
 error: called `filter(..).next_back()` on an `DoubleEndedIterator`. This is more succinctly expressed by calling `.rfind(..)` instead
-  --> tests/ui/methods_fixable.rs:18:13
+  --> tests/ui/filter_next.rs:43:13
    |
 LL |     let _ = vec![1].into_iter().filter(|&x| x < 0).next_back();
    |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `vec![1].into_iter().rfind(|&x| x < 0)`
