--- tests/ui/methods_unfixable.stderr
@@ -1,16 +1,0 @@
-error: called `filter(..).next()` on an `Iterator`. This is more succinctly expressed by calling `.find(..)` instead
-  --> tests/ui/methods_unfixable.rs:7:13
-   |
-LL |     let _ = iter.filter(|_| true).next();
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `iter.find(|_| true)`
-   |
-help: you will also need to make `iter` mutable, because `find` takes `&mut self`
-  --> tests/ui/methods_unfixable.rs:6:9
-   |
-LL |     let iter = (0..10);
-   |         ^^^^
-   = note: `-D clippy::filter-next` implied by `-D warnings`
-   = help: to override `-D warnings` add `#[allow(clippy::filter_next)]`
-
-error: aborting due to 1 previous error
-
